My last commit.
Added cells and pages for Indexes that "should" work, but I didn't bother to create any tests for them. The IndexTree is largely unfinished.

Changed sequential search over to binary search when searching within Pages for data cells.

Good job guys thanks for all of your hard work. This was a beast of a project and and I'm sure had this been longer than a 2 weeks we would have been able to do it.